Fix build failures in Scala 2.11.12
